Sped up receive_line function
Instead of receiving one char at a time until reaching the end of the line, now a buffer to store several data is used. If the buffer does not contain full lines than other data are received from the socket, otherwise the first line in the buffer is returned and the buffer is updated removing the returned line.